refactor: Member 패키지 내 Entity를 repository.entity 패키지 내로 이동
