+++ src/main/java/com/bookbla/americano/domain/member/repository/MemberRepository.java
@@ -2,7 +2,7 @@ package com.bookbla.americano.domain.member.repository;
 
 
 import com.bookbla.americano.base.exception.BaseException;
-import com.bookbla.americano.domain.member.Member;
+import com.bookbla.americano.domain.member.repository.entity.Member;
 import com.bookbla.americano.domain.member.enums.MemberType;
 import com.bookbla.americano.domain.member.exception.MemberExceptionType;
 import java.util.Optional;
