+++ src/main/java/com/bookbla/americano/domain/member/repository/MemberStyleRepository.java
@@ -1,8 +1,8 @@
 package com.bookbla.americano.domain.member.repository;
 
 import com.bookbla.americano.base.exception.BaseException;
-import com.bookbla.americano.domain.member.Member;
-import com.bookbla.americano.domain.member.MemberStyle;
+import com.bookbla.americano.domain.member.repository.entity.Member;
+import com.bookbla.americano.domain.member.repository.entity.MemberStyle;
 import com.bookbla.americano.domain.member.exception.MemberExceptionType;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
