+++ src/main/java/com/bookbla/americano/domain/memberask/repository/MemberAskRepository.java
@@ -1,7 +1,7 @@
 package com.bookbla.americano.domain.memberask.repository;
 
 import com.bookbla.americano.base.exception.BaseException;
-import com.bookbla.americano.domain.member.Member;
+import com.bookbla.americano.domain.member.repository.entity.Member;
 import com.bookbla.americano.domain.memberask.exception.MemberAskExceptionType;
 import com.bookbla.americano.domain.memberask.repository.entity.MemberAsk;
 import java.util.Optional;
